refactor: 구현코드 이동
