+++ src/main/kotlin/lotto/LottoTicket.kt
@@ -1,0 +1,5 @@
+package lotto
+
+class LottoTicket(val count: Int) {
+    constructor(money: Money) : this(money / TicketAmount().money)
+}
