+++ src/main/kotlin/lotto/Match.kt
@@ -1,0 +1,5 @@
+package lotto
+
+class Match(left: List<Int>, right: List<Int>) {
+    val count: Int = left.intersect(right).size
+}
