+++ src/main/kotlin/lotto/TicketAmount.kt
@@ -1,0 +1,6 @@
+package lotto
+
+class TicketAmount(count: Int = 1) {
+    val money: Money = Money(1_000 * count)
+    val amount: Int = money.amount
+}
