+++ src/test/kotlin/lotto/LottoTicketTest.kt
@@ -19,22 +19,4 @@ class LottoTicketTest {
     fun `나머지를 제외한 몫만 제공한다`() {
         assertThat(Money(1_500) / Money(1_000)).isEqualTo(1)
     }
-
-    class LottoTicket(val count: Int) {
-        constructor(money: Money) : this(money / TicketAmount().money)
-    }
-
-    data class Money(val amount: Int) {
-        operator fun div(divisor: Money): Int {
-            require(divisor.amount != 0)
-            return amount / divisor.amount
-        }
-
-        operator fun times(other: Int): Int = amount * other
-    }
-
-    class TicketAmount(count: Int = 1) {
-        val money: Money = Money(1_000 * count)
-        val amount: Int = money.amount
-    }
 }
