+++ src/test/kotlin/lotto/RankingTest.kt
@@ -1,6 +1,5 @@
 package lotto
 
-import lotto.LottoDrawMachineTest.LottoNumber
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
@@ -23,30 +22,5 @@ class RankingTest {
         assertThat(
             Ranking(winningNumber, LottoNumber(pickNumbers)).rank
         ).isEqualTo(rank)
-    }
-
-    class Ranking(match: Match) {
-        val rank: Rank = Rank.of(match.count)
-
-        constructor(theNumber: LottoNumber, other: LottoNumber) : this(Match(theNumber, other))
-
-        enum class Rank(private val matchCount: Int = 0, private val amount: Long = 0) {
-            FIRST(6, 2_000_000_000L),
-            SECOND(5, 1_500_000L),
-            THIRD(4, 50_000L),
-            FOURTH(3, 5_000L),
-            MISS;
-
-            fun prize(count: Int): Long = amount * count
-            private fun same(count: Int): Boolean = matchCount == count
-
-            companion object {
-                fun of(count: Int): Rank = values().firstOrNull { it.same(count) } ?: MISS
-            }
-        }
-    }
-
-    class Match(left: List<Int>, right: List<Int>) {
-        val count: Int = left.intersect(right).size
     }
 }
